+++ src/asp/Core/StereoSettings.cc
@@ -444,6 +444,8 @@ namespace asp {
                             "The width and height of the GUI window in pixels.")
       ("single-window,w",   po::bool_switch(&global.single_window)->default_value(false)->implicit_value(true),
                             "Show all images in the same window (with a dialog to choose among them) rather than next to each other.")
+      ("view-several-side-by-side",   po::bool_switch(&global.view_several_side_by_side)->default_value(false)->implicit_value(true),
+                            "View several images side-by-side, with a dialog to choose which images to show.")
       ("use-georef,g",        po::bool_switch(&global.use_georef)->default_value(false)->implicit_value(true),
                             "Plot the images in the projected coordinate system given by the georeference of the images.")
       ("hillshade",         po::bool_switch(&global.hillshade)->default_value(false)->implicit_value(true),