+++ src/asp/Core/StereoSettings.h
@@ -285,7 +285,7 @@ namespace asp {
     bool hillshade;
     int lowest_resolution_subimage_num_pixels;
     double hillshade_azimuth, hillshade_elevation;
-    bool view_matches, colorize, preview;
+    bool view_matches, view_several_side_by_side, colorize, preview;
     std::string match_file, gcp_file, dem_file, csv_datum, csv_format_str, csv_proj4, nvm;
     bool delete_temporary_files_on_exit;
     bool create_image_pyramids_only, hide_all;